refactor: export ai actions utilities
